% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_codelist}
\alias{get_codelist}
\title{Get Codelist}
\usage{
get_codelist(dataset_id, values = "levels")
}
\arguments{
\item{dataset_id}{The dataset ID of the dataflow you wish to get a codelist for. 
Dataset IDs can be found using get_dataflows().}

\item{values}{Should the returned codelist include only variables ("vars"), 
or variables and variable levels ("levels"). Default is set to return both variables and levels.}
}
\value{
A tibble with the available variables names, types and levels.
}
\description{
Returns the codelist of variables and variable levels for the selected dataflow.
}
\examples{
# Return the Codelist for the Apparent Consumption of Alcohol dataflow (dataset_id = "ALC")
codes <- get_codelist(dataset_id = "ALC");

# Only retrieve variable names
var_codes <- get_codelist(dataset_id = "ALC", values = "vars");
}
