% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(
  dataset_id,
  start_date = NA_real_,
  end_date = NA_real_,
  add_labels = TRUE,
  filters = NULL
)
}
\arguments{
\item{dataset_id}{The dataset ID of the dataflow you wish to get a codelist for. 
Dataset IDs can be found using get_dataflows().}

\item{start_date}{Optional parameter, refines the search to only include data after the provided start date (to be implemented).}

\item{end_date}{Optional parameter, refines the search to only include data before the provided end date (to be implemented).}

\item{add_labels}{Should additional variable level label columns be included in the output. Default is TRUE.}

\item{filters}{Additional filter for other dimensions of the data (to be implemented).}
}
\value{
A tibble with the data from the dataset.
}
\description{
Returns the dataset for the selected dataflow.
}
\examples{
# Return the Codelist for the Apparent Consumption of Alcohol dataflow (dataset_id = "ALC")
codes <- get_codelist(dataset_id = "ALC");

# Only retrieve variable names
var_codes <- get_codelist(values = "vars");
}
