% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_dataflows}
\alias{get_dataflows}
\title{Get Dataflows}
\usage{
get_dataflows(search = ".", case_sensitive = FALSE, exact = FALSE)
}
\arguments{
\item{search}{Search term (regular expression). Default returns all dataflows.}

\item{case_sensitive}{Should the search term be treated as case sensitive, default set to FALSE.}

\item{exact}{Return only exact matches of the search term, default set to FALSE}
}
\value{
A tibble containing information on the matched dataflows, including their dataset_id, name and description.
}
\description{
Returns dataflows currently available from the ABS data API which match the supplied search term.
}
\examples{
# Return all available dataflows
all_data <- get_dataflows();

# Return all dataflows which mention the term WPI
wage_price_index <- get_dataflows(search = "wpi");

# Return all dataflows which begin with WPI
wage_price_index <- get_dataflows(search = "^WPI", case_sensitive = TRUE);

# Return all datasets related to indigenous/aboriginal/torres strait islander Australians.
indigenous_data <- get_dataflows(search = "indigenous|aboriginal|torres strait");
}
