% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(
  dataset_id,
  level_type = "both",
  start_date = NA_real_,
  end_date = NA_real_,
  filters = NULL
)
}
\arguments{
\item{dataset_id}{The dataset ID of the dataflow you wish to get a codelist for. 
Dataset IDs can be found using get_dataflows().}

\item{level_type}{Select variable level return type from "codes", "labels" or "both". Default is "both".}

\item{start_date}{Optional parameter, refines the search to only include data after the provided start date (to be implemented).}

\item{end_date}{Optional parameter, refines the search to only include data before the provided end date (to be implemented).}

\item{filters}{Additional filter for other dimensions of the data (to be implemented).}
}
\value{
A tibble with the data from the dataset.
}
\description{
Returns the dataset for the selected dataflow.
}
\examples{
# Return the dataset for the Apparent Consumption of Alcohol dataflow (dataset_id = "ALC")
data <- get_data(dataset_id = "ALC");

# Include only variable names
data_labelled <- get_data(dataset_id = "ALC", level_type = "labels");
}
